# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "description": "API для управления задачами",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить список задач",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать задачу",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданная задача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить одну задачу",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали задачи",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленная задача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Задача удалена"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "properties": {
                    "id": {
                        "description": "ID задачи",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Заголовок задачи",
                        "type": "string"
                    },
                    "description": {
                        "description": "Описание задачи",
                        "type": "string"
                    },
                    "status": {
                        "description": "Статус задачи (завершена или нет)",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Дата создания задачи",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления задачи",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TaskRequest": {
                "properties": {
                    "title": {
                        "description": "Заголовок задачи",
                        "type": "string",
                        "example": "New Task"
                    },
                    "description": {
                        "description": "Описание задачи",
                        "type": "string",
                        "example": "Description of the task"
                    },
                    "status": {
                        "description": "Статус задачи",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "API для работы с задачами"
        }
    ]
}